Read response prompt after sending imagine
Added an extra step to the preview process where we read the response
message after the imagine command has been sent. This way we can get
the modified prompt and use it to match following messages.

This fixes the issue where prompts with certain links would not be
matched.

This change has been added only to midjourney.
Bluewillow doesn't send a response prompt after the imagine command.