Handle midjourney queue errors
When job queued error is received instead of returning an error the
response is read from the embedded data of the message and the process
continues.

When a non temporary error is received such as a "queue full" error
the tool should wait before retrying to avoid flooding the queue.

References #14