Fix signed discord CDN URLs
Discord CDN URLs have a signature at the end of the url that can not be
removed. This commit removes the cleaning of the url and adjust the
file name generation to be compatible with these URLs.

This change on discord CDN URLs was announced some time ago, but it was
not affecting URLs until now. This change also makes the URLs expire
after a while, so it is recommended to download the images.

An alternative to this would be to use midjourney CDN URLs instead of
discord CDN URLs.

Closes #78